requirements.txt
asttokens==2.4.0
av==10.0.0
backcall==0.2.0
backports.functools-lru-cache==1.6.5
certifi==2024.8.30
colorama==0.4.6
comm==0.1.4
contourpy==1.0.5
cycler==0.11.0
debugpy==1.6.7
decorator==5.1.1
exceptiongroup==1.1.3
executing==1.2.0
ffpyplayer==4.5.0
fonttools==4.25.0
importlib-metadata==6.8.0
imutils==0.5.4
ipykernel==6.25.2
ipython==8.15.0
jedi==0.19.0
joblib==1.2.0
jupyter_client==8.3.1
jupyter_core==5.3.1
kiwisolver==1.4.4
matplotlib==3.7.2
matplotlib-inline==0.1.6
mkl-fft==1.3.6
mkl-random==1.2.2
mkl-service==2.4.0
munkres==1.1.4
nest-asyncio==1.5.6
numpy==1.24.2
opencv-python==4.7.0.72
packaging==23.0
pandas==2.1.1
parso==0.8.3
pickleshare==0.7.5
Pillow==9.4.0
pip==23.0.1
platformdirs==3.10.0
ply==3.11
prompt-toolkit==3.0.39
psutil==5.9.0
pure-eval==0.2.2
PyAudio==0.2.13
Pygments==2.16.1
pyparsing==3.0.9
PyQt5==5.15.7
PyQt5-sip==12.11.0
python-dateutil==2.8.2
pytz==2023.3.post1
pywin32==305.1
pyzmq==23.2.1
scikit-learn==1.2.2
scipy==1.10.1
setuptools==65.6.3
sip==6.6.2
six==1.16.0
stack-data==0.6.2
threadpoolctl==3.1.0
tk==0.1.0
tkvideoplayer==2.3
toml==0.10.2
tornado==6.3.2
traitlets==5.10.0
typing_extensions==4.7.1
tzdata==2023.3
wcwidth==0.2.6
wheel==0.38.4
wincertstore==0.2
zipp==3.16.2
